REF: set_def_signal.c set_interactive.cを作成
