+++ signal/includes/sigaction.h
@@ -6,7 +6,7 @@
 /*   By: tkirihar <tkirihar@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/11 23:34:51 by tkirihar          #+#    #+#             */
-/*   Updated: 2022/02/13 01:22:57 by tkirihar         ###   ########.fr       */
+/*   Updated: 2022/02/13 16:41:19 by tkirihar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,13 +25,13 @@
 
 int	g_exit_status;
 
-// set_signal.c
+// set_dfl_signal.c
 
-void	set_interactive_sigint(void);
-void	set_dfl_sigint(void);
-void	set_dfl_sigquit(void);
-void	set_interactive_signal(void);
 void	set_dfl_signal(void);
+
+// set_interactive_signal.c
+
+void	set_interactive_signal(void);
 
 // sigaction.c
